I suppose that the expression argument of VCAS can be hidden; automatic derivation seems to work pretty well.
The worst case scenario is needing to revert this commit.

Well, actually, the worst case scenario is needing to scrap everything and start over.

Well, actually, the *worst* case scenario is having a crappy, unverified version of Casanova Fly be used in an important setting and start WWIII or something.  *That* would be *bad*.

Well, actually...