The algebraic equality type is once again a field value.
I made this change primarily because the previous, programmatically-derived algebraic equality type lacked the ability to cope with commutativity; the previous type would not recognize that \(a + b\) is *always* algebraically equal to \(b + a\)... well, if _+_ is commutative, anyway.  :-)

The main costs of this change are a different universe level and losing some progress on equality proofs.